Add support for non-compliant mipmaps that report 0 pitch
This tool uses the libsquish DecompressImage function signature with the
explicit pitch size, but there also exists a signature where it is
unspecified and defaulted to width * 4. If another tool (likely Matt's
DST Tools) specifies pitch as 0 (essentially ignores it) and only uses
the CompressImage and DecompressImage functions from libsquish where a
pitch is not specified, then those TEX files would work properly with
said tools but break the tools that use the pitch explicit functions
(i.e. this one).

To handle this case, if the meta-data of a mipmap has a pitch of 0, the
pitch is then set to the mipmap's with times 4, in accordance with
libsquish's handling of implicit pitch, while being read